move the middleware function back into the lib files, track only patchwork in egress
