+++ ingestion/src/lib.rs
@@ -1,7 +1,7 @@
 use axum::{
-    extract::{FromRef, State},
-    middleware,
-    response::Json,
+    extract::{FromRef, MatchedPath, Request, State},
+    middleware::{self, Next},
+    response::{IntoResponse, Json},
     routing::post,
     Router,
 };
@@ -18,7 +18,7 @@ use tracing::{error, info};
 pub mod cron;
 pub mod legacy;
 pub mod util;
-use ::util::{middleware::track_request_duration, DbPools, PooledPgConn};
+use ::util::{DbPools, PooledPgConn};
 use util::{
     levels::{self, LevelTable},
     permissions::{self, PermitTables},
@@ -498,6 +498,28 @@ pub fn get_conversions(filename: &str) -> Result<ParamConversions, Error> {
             })
             .collect::<Result<HashMap<String, ReferenceParam>, csv::Error>>()?,
     ))
+}
+
+/// Middleware function that runs around a request, so we can record how long it took
+async fn track_request_duration(req: Request, next: Next) -> impl IntoResponse {
+    let start = std::time::Instant::now();
+    let path = if let Some(matched_path) = req.extensions().get::<MatchedPath>() {
+        matched_path.as_str().to_owned()
+    } else {
+        req.uri().path().to_owned()
+    };
+    let method = req.method().to_string();
+
+    let response = next.run(req).await;
+
+    let duration = start.elapsed().as_secs_f64();
+    let status = response.status().as_u16().to_string();
+
+    let labels = [("method", method), ("path", path), ("status", status)];
+
+    metrics::histogram!("http_requests_duration_seconds", &labels).record(duration);
+
+    response
 }
 
 pub async fn run(
