+++ util/src/lib.rs
@@ -11,7 +11,6 @@ use tokio_util::sync::CancellationToken;
 pub mod deserialize;
 pub mod dut_parse;
 pub mod idf_parse;
-pub mod middleware;
 
 pub type PooledPgConn<'a> = PooledConnection<'a, PostgresConnectionManager<NoTls>>;
 pub type PgPool = bb8::Pool<PostgresConnectionManager<NoTls>>;
